refactor: Move types to types.ts
