+++ src/index.ts
@@ -1,7 +1,7 @@
 import pino from "pino";
 import type Pulsar from "pulsar-client";
 import dotenv = require("dotenv");
-import { getConfig } from "./config";
+import getConfig from "./config";
 import createHealthCheckServer from "./healthCheck";
 import keepProcessingMessages from "./messageProcessing";
 import {
